refactor: update CLI module to improve separation of concerns
Temporarily using --no-verify as rustfmt configuration in pre-commit hooks has issues with import ordering that can't be resolved with standard rustfmt. A TODO item has been added to fix the formatting issues with rustfmt configuration.
